clean up previous POC
