+++ data/collector.json
@@ -1,1 +1,1 @@
-[{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}}]
+[{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055614}},{"dogecoin":{"usd":0.055467}},{"dogecoin":{"usd":0.055467}},{"dogecoin":{"usd":0.055467}},{"dogecoin":{"usd":0.055467}},{"dogecoin":{"usd":0.055467}},{"dogecoin":{"usd":0.055467}},{"dogecoin":{"usd":0.055467}},{"dogecoin":{"usd":0.055467}}]
